rework inheritance map
